Fixes bug in dry-run for kpt live migrate

+++ commands/migratecmd.go
@@ -234,7 +234,11 @@ func (mr *MigrateRunner) retrieveInvObjs(invObj inventory.InventoryInfo) ([]obje
 func (mr *MigrateRunner) migrateObjs(cmObjs []object.ObjMetadata, reader io.Reader, args []string) error {
 	fmt.Fprint(mr.ioStreams.Out, "  migrate inventory to ResourceGroup...")
 	if len(cmObjs) == 0 {
-		fmt.Fprint(mr.ioStreams.Out, "no inventory objects found\n")
+		fmt.Fprintln(mr.ioStreams.Out, "no inventory objects found")
+		return nil
+	}
+	if mr.dryRun {
+		fmt.Fprintln(mr.ioStreams.Out, "success")
 		return nil
 	}
 	rgReader, err := mr.rgLoader.ManifestReader(reader, args)
@@ -259,7 +263,7 @@ func (mr *MigrateRunner) migrateObjs(cmObjs []object.ObjMetadata, reader io.Read
 	if err != nil {
 		return err
 	}
-	fmt.Fprint(mr.ioStreams.Out, "success\n")
+	fmt.Fprintln(mr.ioStreams.Out, "success")
 	return nil
 }
 
@@ -270,6 +274,9 @@ func (mr *MigrateRunner) deleteConfigMapInv(invObj inventory.InventoryInfo) erro
 	cmInvClient, err := mr.cmProvider.InventoryClient()
 	if err != nil {
 		return err
+	}
+	if mr.dryRun {
+		cmInvClient.SetDryRunStrategy(common.DryRunClient)
 	}
 	if err = cmInvClient.DeleteInventoryObj(invObj); err != nil {
 		return err
@@ -292,12 +299,14 @@ func (mr *MigrateRunner) deleteConfigMapFile() error {
 		}
 		if len(cmFilename) > 0 {
 			fmt.Fprintf(mr.ioStreams.Out, "deleting inventory template file: %s...", cmFilename)
-			err = os.Remove(cmFilename)
-			if err != nil {
-				fmt.Fprint(mr.ioStreams.Out, "failed\n")
-				return err
+			if !mr.dryRun {
+				err = os.Remove(cmFilename)
+				if err != nil {
+					fmt.Fprintln(mr.ioStreams.Out, "failed")
+					return err
+				}
 			}
-			fmt.Fprint(mr.ioStreams.Out, "success\n")
+			fmt.Fprintln(mr.ioStreams.Out, "success")
 		}
 	}
 	return nil
